+++ Mercadona/Models/PromotionsProduitsModel.cs
@@ -14,7 +14,7 @@
         public int Id { get; set; }
         public int IdProduit { get; set; }
         public int IdPromotion { get; set; }
-        public int prixPromo { get; set; }
+        public float prixPromo { get; set; }
 
         public IEnumerable<ProduitsModel> Produits { get; set; }
         public IEnumerable<PromotionsModel> Promotions { get; set;}
